Update README.md
In this project, I manipulated the DOM to created images, then manipulate those images to have them move in the X and Y coordinates. They also bounced off borders of the window and reversed direction. An HTML web page was created for the user to click a button to add more PacMen. You can click another button to start the random movement.